Examples/MessageBusComm/requirements.txt
# Kafka dependencies
kafka-python==2.0.2

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# FastAPI dependencies
fastapi==0.103.2
uvicorn==0.23.2

# Utility dependencies
python-dotenv==1.0.0
pydantic==2.4.2
